+++ packages/framework/tests/Feature/AutomaticNavigationConfigurationsTest.php
@@ -457,28 +457,23 @@ class AutomaticNavigationConfigurationsTest extends TestCase
         ]);
     }
 
-    public function testMainNavigationMenuItemsWithTheSameLabelAreFiltered()
+    public function testMainNavigationMenuItemsWithTheSameLabelAreNotFilteredForDuplicates()
     {
-        $this->assertMenuEquals(['Foo'], [
-            new MarkdownPage('foo'),
-            new MarkdownPage('foo'),
-        ]);
-
-        $this->assertMenuEquals(['Foo'], [
+        $this->assertMenuEquals(['Foo', 'Foo'], [
             new MarkdownPage('foo', ['navigation.label' => 'Foo']),
             new MarkdownPage('bar', ['navigation.label' => 'Foo']),
         ]);
     }
 
-    public function testMainNavigationMenuItemsWithTheSameLabelAreFilteredCaseInsensitive()
+    public function testMainNavigationMenuItemsWithTheSameLabelAreNotFilteredForDuplicatesRegardlessOfCase()
     {
-        $this->assertMenuEquals(['Foo'], [
+        $this->assertMenuEquals(['Foo', 'Foo', 'FOO'], [
             new MarkdownPage('foo'),
             new MarkdownPage('Foo'),
             new MarkdownPage('FOO'),
         ]);
 
-        $this->assertMenuEquals(['foo'], [
+        $this->assertMenuEquals(['foo', 'Foo', 'FOO'], [
             new MarkdownPage('foo', ['navigation.label' => 'foo']),
             new MarkdownPage('bar', ['navigation.label' => 'Foo']),
             new MarkdownPage('baz', ['navigation.label' => 'FOO']),
@@ -882,28 +877,23 @@ class AutomaticNavigationConfigurationsTest extends TestCase
         }
     }
 
-    public function testSidebarItemsWithTheSameLabelAreFiltered()
+    public function testSidebarItemsWithTheSameLabelAreNotFiltered()
     {
-        $this->assertSidebarEquals(['Foo'], [
-            new DocumentationPage('foo'),
-            new DocumentationPage('foo'),
-        ]);
-
-        $this->assertSidebarEquals(['Foo'], [
+        $this->assertSidebarEquals(['Foo', 'Foo'], [
             new DocumentationPage('foo', ['navigation.label' => 'Foo']),
             new DocumentationPage('bar', ['navigation.label' => 'Foo']),
         ]);
     }
 
-    public function testSidebarItemsWithTheSameLabelAreFilteredCaseInsensitive()
+    public function testSidebarItemsWithTheSameLabelAreNotFilteredForDuplicatesRegardlessOfCase()
     {
-        $this->assertSidebarEquals(['Foo'], [
+        $this->assertSidebarEquals(['Foo', 'Foo', 'FOO'], [
             new DocumentationPage('foo'),
             new DocumentationPage('Foo'),
             new DocumentationPage('FOO'),
         ]);
 
-        $this->assertSidebarEquals(['foo'], [
+        $this->assertSidebarEquals(['foo', 'Foo', 'FOO'], [
             new DocumentationPage('foo', ['navigation.label' => 'foo']),
             new DocumentationPage('bar', ['navigation.label' => 'Foo']),
             new DocumentationPage('baz', ['navigation.label' => 'FOO']),